replace mobile upload drawer
