+++ packages/common/src/store/ui/modals/types.ts
@@ -17,7 +17,6 @@ export type Modals =
   | 'FeedFilter'
   | 'TrendingGenreSelection'
   | 'SocialProof'
-  | 'MobileUpload'
   | 'EditFolder'
   | 'SignOutConfirmation'
   | 'Overflow'
